training/requirements.txt
# Core ML
torch>=2.1.0
transformers>=4.36.0
datasets>=2.16.0
accelerate>=0.25.0

# Experiment Tracking
mlflow>=2.10.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Active Learning
modAL-python>=0.4.1

# Utilities
tqdm>=4.66.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Optional: Data Labeling
# label-studio>=1.10.0
# prodigy>=1.12.0
